Fix missing Blue's Journey instruments
Blue's Journey vroma0 file is 3MB.  It appears the first MB is for ADPCMA, the middle 2MB appears to be blank (and is currently being written to ADPCMA memory space after the first MB; this doesn't appear to cause issues). The third MB looks like it is for ADPCMB.  The code appears to reference  samples at 0x1XXXXX. This requires mirroring to work.