ADPCM-A: Set done only when channel is on.
It was possible for the done signal to go high while the channel was off
which made the channel end bit in port 6 get set.

Result: missing sample because the Z80 sees end bit set immediately
after sending A-ON and stops the channel.

Caused by changing the end address and then changing it back to the
previous end address.